refactor: moved rendering helpers to rendering/utils folder
