+++ src/rendering/utils/checkPoints.js
