+++ src/rendering/utils/createPath.js
