+++ src/rendering/utils/createSVGStyle.js
